requirements.txt
flask
tensorflow
numpy
pandas
opencv-python
Pillow